% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_client_api_token.R
\name{get_client_api_token}
\alias{get_client_api_token}
\title{get_client_api_token}
\usage{
get_client_api_token(client_id, client_secret, domain)
}
\arguments{
\item{client_id}{Auth0 app client id}

\item{client_secret}{Auth0 app secret}

\item{Domain}{Auth0 app domain}
}
\value{
Returns full api token response
}
\description{
Get app api token used to read app profile
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 api_tokens<-get_client_api_token(app_client_id,app_secret,Domain)

client<-  get_client_profile(api_tokens$access_token,app_client_id,Domain)

}
}
}
\seealso{
\code{\link[httr]{POST}},\code{\link[httr]{content}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
}
