% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_client_profile.R
\name{get_app_client_profile}
\alias{get_app_client_profile}
\title{get_app_client_profile}
\usage{
get_app_client_profile(access_token, client_id, domain)
}
\arguments{
\item{access_token}{Auth0 app api access token.}

\item{domain}{Auth0 domain.}

\item{app_client_id}{app client id.}
}
\value{
returns client response for app.
}
\description{
Get app client profile.  Used to read app metadata.
}
\details{
Details
}
\examples{
\dontrun{
if(interactive(){

 api_tokens<-ShinyPlatform::get_app_api_token(app_client_id,app_secret,domain)
 client<-  ShinyPlatform::get_app_client_profile(api_tokens$access_token,app_client_id,domain)

}
}
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[httr]{GET}},\code{\link[httr]{content}}
}
