% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_profile.R
\name{get_user_profile}
\alias{get_user_profile}
\title{get_user_profile}
\usage{
get_user_profile(access_token, id_token, domain)
}
\arguments{
\item{access_token}{Auth0 access token}

\item{id_token}{Auth0 id token}

\item{domain}{auth0 Domain}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get Auth0 user profile
}
\details{
return Auth0 user profile
}
\examples{
\dontrun{
if(interactive()){
  profile <-  ShinyPlatform::get_user_profile(token$access_token,token$id_token,domain)
 }
}
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[jose]{base64url_encode}}
 \code{\link[httr]{GET}},\code{\link[httr]{content}}
}
