% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_auth_code.R
\name{has_auth_code}
\alias{has_auth_code}
\title{has_auth_code}
\usage{
has_auth_code(params)
}
\arguments{
\item{params}{session URL}
}
\value{
Tru eor False
}
\description{
checks for authorization code in url
}
\details{
params is a list object containing the parsed URL parameters. Return TRUE if
based on these parameters, it looks like auth codes are present that we can
use to get an access token. If not, it means we need to go through the OAuth
flow.
}
\examples{
\dontrun{
if(interactive()){
 has_auth_code(parseQueryString(req$QUERY_STRING))
 }
}
}
